requirements.txt
addict==2.4.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.38.0
idna==3.4
importlib-metadata==5.2.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.6.2
mmcv-full==1.7.0
model-index==0.1.11
numpy==1.24.0
opencv-python==4.6.0.66
openmim==0.3.3
ordered-set==4.1.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pycocotools==2.0.6
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
rich==12.6.0
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
terminaltables==3.1.10
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.4.0
urllib3==1.26.13
yapf==0.32.0
zipp==3.11.0
pyserial>=3.1.1
Pymodbus>=2.5.3
